✨ Start server in background on page load.
When the page is loaded, send a ping (GET /) request in a background
thread. When the photo is uploaded and the ping request has not
completed yet, inform that the service is starting up.

Once the ping completed successfully, send the predict request.

Also, change error handling in the predict function, return exception
instead of raising.